refactor: move business logic from controller to service layer

+++ target30k-api/src/main/java/br/com/jsctecnologia/target30k_api/model/FreelaEarning.java
@@ -2,6 +2,8 @@ package br.com.jsctecnologia.target30k_api.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -14,6 +16,6 @@ public class FreelaEarning {
 
     private String description;
     @Column(name = "valor")
-    private Double value;
+    private BigDecimal value;
     private LocalDate date;
 }
